backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.2.3
python-dotenv==1.0.1
twilio==9.3.2
